issue-1444: move netlink helpers to core/libs/netlink
