+++ cloud/storage/core/libs/netlink/netlink_socket.cpp
@@ -1,0 +1,98 @@
+#include "netlink_socket.h"
+
+#include <netlink/genl/ctrl.h>
+#include <netlink/genl/genl.h>
+
+namespace NCloud::NNetlink {
+
+namespace {
+
+using TResponseHandler = std::function<int(nl_msg*)>;
+
+class TNetlinkSocket: public INetlinkSocket
+{
+private:
+    nl_sock* Socket;
+    int Family;
+
+public:
+    explicit TNetlinkSocket(TString netlinkFamily)
+    {
+        Socket = nl_socket_alloc();
+
+        if (Socket == nullptr) {
+            throw TServiceError(E_FAIL) << "unable to allocate netlink socket";
+        }
+
+        if (int err = genl_connect(Socket)) {
+            nl_socket_free(Socket);
+            throw TServiceError(E_FAIL)
+                << "unable to connect to generic netlink socket: "
+                << nl_geterror(err);
+        }
+
+        Family = genl_ctrl_resolve(Socket, netlinkFamily.c_str());
+
+        if (Family < 0) {
+            nl_socket_free(Socket);
+            throw TServiceError(E_FAIL)
+                << "unable to resolve netlink family: " << nl_geterror(Family);
+        }
+    }
+
+    ~TNetlinkSocket() override
+    {
+        nl_socket_free(Socket);
+    }
+
+    [[nodiscard]] int GetFamily() const override
+    {
+        return Family;
+    }
+
+    void SetCallback(nl_cb_type type, TNetlinkSocketCallback func) override
+    {
+        auto arg = std::make_unique<TResponseHandler>(std::move(func));
+
+        if (int err = nl_socket_modify_cb(
+                Socket,
+                type,
+                NL_CB_CUSTOM,
+                TNetlinkSocket::ResponseHandler,
+                arg.get()))
+        {
+            throw TServiceError(E_FAIL)
+                << "unable to set socket callback: " << nl_geterror(err);
+        }
+        arg.release();
+    }
+
+    static int ResponseHandler(nl_msg* msg, void* arg)
+    {
+        auto func = std::unique_ptr<TResponseHandler>(
+            static_cast<TResponseHandler*>(arg));
+
+        return (*func)(msg);
+    }
+
+    void Send(nl_msg* message) override
+    {
+        if (int err = nl_send_auto(Socket, message); err < 0) {
+            throw TServiceError(E_FAIL) << "send error: " << nl_geterror(err);
+        }
+        if (int err = nl_wait_for_ack(Socket)) {
+            // this is either recv error, or an actual error message received
+            // from the kernel
+            throw TServiceError(E_FAIL) << "recv error: " << nl_geterror(err);
+        }
+    }
+};
+
+}   // namespace
+
+INetlinkSocketPtr CreateNetlinkSocket(TString netlinkFamily)
+{
+    return std::make_unique<TNetlinkSocket>(std::move(netlinkFamily));
+}
+
+}   // namespace NCloud::NNetlink
